+++ packages/controls/js/libs/box-spacing-control/components/margin/margin-right.js
@@ -50,6 +50,7 @@ export function MarginRight({
 	});
 
 	const [labelClassName, setLabelClassName] = useState('');
+	const [mouseDownTime, setMouseDownTime] = useState(0);
 
 	// $FlowFixMe
 	const { isSetValueAddon, ValueAddonPointer, valueAddonControlProps } =
@@ -160,11 +161,16 @@ export function MarginRight({
 									return;
 								}
 
-								if (sideSpace.unit === 'func') {
+								if (
+									_isSetValueAddon ||
+									sideSpace?.unit === 'func'
+								) {
 									event.preventDefault();
 									return;
 								}
 
+								// Store the mouse down timestamp
+								setMouseDownTime(Date.now());
 								onDragStart(event);
 								setFocusSide(sideId);
 							},
@@ -188,13 +194,15 @@ export function MarginRight({
 					}
 				}}
 				onClick={(event) => {
-					// open on double click
-					// or value addon
-					// or CSS Value
+					// Calculate click duration
+					const clickDuration = Date.now() - mouseDownTime;
+
+					// If it's a quick click (less than 200ms) or it's a value addon/func
 					if (
 						_isSetValueAddon ||
 						sideSpace?.unit === 'func' ||
-						event.detail > 1
+						event.detail > 1 ||
+						clickDuration < 200
 					) {
 						setFocusSide(sideId);
 						setOpenPopover(sideId);
@@ -241,6 +249,16 @@ export function MarginRight({
 
 						{openPopover === sideId && (
 							<SidePopover
+								hasValue={value?.margin?.right}
+								resetToDefault={() => {
+									setValue({
+										...value,
+										margin: {
+											...value.margin,
+											right: '',
+										},
+									});
+								}}
 								defaultValue={defaultValue}
 								id={getId(id, 'margin.right')}
 								offset={255}