+++ packages/controls/js/libs/box-spacing-control/components/padding/padding-all.js
@@ -49,6 +49,7 @@ export function PaddingAll({
 	});
 
 	const [labelClassName, setLabelClassName] = useState('');
+	const [mouseDownTime, setMouseDownTime] = useState(0);
 
 	// $FlowFixMe
 	const { isSetValueAddon, ValueAddonPointer, valueAddonControlProps } =
@@ -153,6 +154,8 @@ export function PaddingAll({
 									return;
 								}
 
+								// Store the mouse down timestamp
+								setMouseDownTime(Date.now());
 								onDragStart(event);
 								setFocusSide(sideId);
 							},
@@ -176,13 +179,15 @@ export function PaddingAll({
 					}
 				}}
 				onClick={(event) => {
-					// open on double click
-					// or value addon
-					// or CSS Value
+					// Calculate click duration
+					const clickDuration = Date.now() - mouseDownTime;
+
+					// If it's a quick click (less than 200ms) or it's a value addon/func
 					if (
 						_isSetValueAddon ||
 						sideSpace?.unit === 'func' ||
-						event.detail > 1
+						event.detail > 1 ||
+						clickDuration < 200
 					) {
 						setFocusSide(sideId);
 						setOpenPopover(sideId);
@@ -328,6 +333,23 @@ export function PaddingAll({
 
 				{openPopover === sideId && (
 					<SidePopover
+						hasValue={
+							value?.padding?.top ||
+							value?.padding?.right ||
+							value?.padding?.bottom ||
+							value?.padding?.left
+						}
+						resetToDefault={() => {
+							setValue({
+								...value,
+								padding: {
+									top: '',
+									right: '',
+									bottom: '',
+									left: '',
+								},
+							});
+						}}
 						defaultValue={defaultValue}
 						id={getId(id, 'padding.top')}
 						type="padding"