+++ packages/controls/js/libs/box-spacing-control/components/padding/padding-bottom.js
@@ -50,6 +50,7 @@ export function PaddingBottom({
 	});
 
 	const [labelClassName, setLabelClassName] = useState('');
+	const [mouseDownTime, setMouseDownTime] = useState(0);
 
 	// $FlowFixMe
 	const { isSetValueAddon, ValueAddonPointer, valueAddonControlProps } =
@@ -169,6 +170,8 @@ export function PaddingBottom({
 									return;
 								}
 
+								// Store the mouse down timestamp
+								setMouseDownTime(Date.now());
 								onDragStart(event);
 								setFocusSide(sideId);
 							},
@@ -192,13 +195,15 @@ export function PaddingBottom({
 					}
 				}}
 				onClick={(event) => {
-					// open on double click
-					// or value addon
-					// or CSS Value
+					// Calculate click duration
+					const clickDuration = Date.now() - mouseDownTime;
+
+					// If it's a quick click (less than 200ms) or it's a value addon/func
 					if (
 						_isSetValueAddon ||
 						sideSpace?.unit === 'func' ||
-						event.detail > 1
+						event.detail > 1 ||
+						clickDuration < 200
 					) {
 						setFocusSide(sideId);
 						setOpenPopover(sideId);
@@ -245,6 +250,16 @@ export function PaddingBottom({
 
 				{openPopover === sideId && (
 					<SidePopover
+						hasValue={value?.padding?.bottom}
+						resetToDefault={() => {
+							setValue({
+								...value,
+								padding: {
+									...value.padding,
+									bottom: '',
+								},
+							});
+						}}
 						defaultValue={defaultValue}
 						id={getId(id, 'padding.bottom')}
 						type="padding"