+++ packages/controls/js/libs/box-spacing-control/components/padding/padding-horizontal.js
@@ -50,6 +50,7 @@ export function PaddingHorizontal({
 	});
 
 	const [labelClassName, setLabelClassName] = useState('');
+	const [mouseDownTime, setMouseDownTime] = useState(0);
 
 	// $FlowFixMe
 	const { isSetValueAddon, ValueAddonPointer, valueAddonControlProps } =
@@ -177,6 +178,8 @@ export function PaddingHorizontal({
 									return;
 								}
 
+								// Store the mouse down timestamp
+								setMouseDownTime(Date.now());
 								onDragStart(event);
 								setFocusSide(sideId);
 							},
@@ -200,13 +203,15 @@ export function PaddingHorizontal({
 					}
 				}}
 				onClick={(event) => {
-					// open on double click
-					// or value addon
-					// or CSS Value
+					// Calculate click duration
+					const clickDuration = Date.now() - mouseDownTime;
+
+					// If it's a quick click (less than 200ms) or it's a value addon/func
 					if (
 						_isSetValueAddon ||
 						sideSpace?.unit === 'func' ||
-						event.detail > 1
+						event.detail > 1 ||
+						clickDuration < 200
 					) {
 						setFocusSide(sideId);
 						setOpenPopover(sideId);
@@ -286,6 +291,17 @@ export function PaddingHorizontal({
 
 				{openPopover === sideId && (
 					<SidePopover
+						hasValue={value?.padding?.left || value?.padding?.right}
+						resetToDefault={() => {
+							setValue({
+								...value,
+								padding: {
+									...value.padding,
+									left: '',
+									right: '',
+								},
+							});
+						}}
 						defaultValue={defaultValue}
 						id={getId(id, 'padding.left')}
 						type="padding"
@@ -294,10 +310,10 @@ export function PaddingHorizontal({
 							setFocusSide('');
 							setOpenPopover('');
 						}}
-						title={__('Left & Right Padding Space', 'blockera')}
+						title={__('Left & Right Padding', 'blockera')}
 						inputLabel={__('Horizontal Padding', 'blockera')}
 						inputLabelPopoverTitle={__(
-							'Horizontal Padding Space',
+							'Horizontal Padding',
 							'blockera'
 						)}
 						inputLabelDescription={