+++ packages/dev-phpunit/php/bootstrap.php
@@ -34,6 +34,7 @@ if (! $_tests_dir) {
 
 if (! file_exists($_tests_dir . '/includes/functions.php')) {
     echo "Could not find $_tests_dir/includes/functions.php, have you run bin/install-wp-tests.sh ?" . PHP_EOL; // WPCS: XSS ok.
+	/* @debug-ignore */
     exit(1);
 }