+++ packages/bootstrap/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+## 1.0.1 (2025-03-29)
+
 ### Improvements
 - Refactored the assets provider to make it easier to extend the plugin.