+++ packages/dev-cypress/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Unreleased
 
-## 1.0.1 (2025-03-15)
+## 1.0.1 (2025-03-29)
 
 ### Improvements
 - Improve Cypress utility functions to work faster and more reliably.