+++ packages/editor/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Unreleased
 
+## 1.2.2 (2025-03-29)
+
+### Automated Tests
+- Added E2E tests for opening popover of spacing feature by clicking on labels or side shapes.
+
 ## 1.2.1 (2025-03-17)
 
 ### Bug Fixes